Move start to __init__
