+++ i3dstatus/service.py
@@ -1,19 +1,14 @@
 """
 Defines and manages d-bus services and objects.
 """
-from gi.repository import GLib
-import json
 import dbus
 import dbus.service
-import sys
-import os
-import yaml
 import weakref
 import threading
-from dbus.mainloop.glib import DBusGMainLoop
 from aioevents import Event
 
 DBUS_SERVICE = 'com.dubstepdish.i3dstatus'
+
 
 class BlockManager(dbus.service.Object):
     """
@@ -49,54 +44,6 @@ class BlockManager(dbus.service.Object):
     def remove_block(self, block):
         pass  # TODO
 
-    @dbus.service.method(INTERFACE, in_signature='a{sv}')
-    def show_block(self, block):
-        # apply config options to the block
-        block_config = {}
-        if block['name'] in self.config:
-            if 'instance' in block and \
-                    block['instance'] in self.config[block['name']]:
-                # this block is configured by instance
-                block_config = self.config[block['name']][block['instance']]
-            else:
-                block_config = self.config[block['name']]
-
-        for i3bar_key in ['color', 'min_width', 'min-width' 'align',
-                          'separator', 'separator_block_width',
-                          'separator-block-width']:
-            if i3bar_key.replace('-', '_') in block:
-                # if given in the block, it overrides the config
-                continue
-            if 'general' in self.config and \
-                    i3bar_key in self.config['general']:
-                # if given in the general config, this is the default for all
-                # blocks
-                k = i3bar_key.replace('-', '_')
-                block[k] = self.config['general'][i3bar_key]
-            if i3bar_key in block_config:
-                # if given in the block config, this overrides the general
-                # config
-                k = i3bar_key.replace('-', '_')
-                block[k] = block_config[i3bar_key]
-
-        # try to find the block given by the name and replace it
-        block_found = False
-        for i, b in enumerate(self.blocks):
-            if b['name'] == block['name']:
-                block_found = True
-                self.blocks[i] = block
-
-        if not block_found:
-            self.blocks.append(block)
-
-        # filter out blocks with no 'full_text' member
-        self.blocks = [b for b in self.blocks if 'full_text' in b and
-                       b['full_text']]
-
-        sys.stdout.write(',' + json.dumps(self.blocks, ensure_ascii=False) +
-                         '\n')
-        sys.stdout.flush()
-
     @dbus.service.method(INTERFACE, in_signature='s', out_signature='v')
     def get_config(self, genname):
         """
@@ -123,12 +70,13 @@ class Block(dbus.service.Object):
         )
 
     changed = Event("Raised when a property is set")
+    # FIXME: Have a D-bus signal of the same name
 
     def __init__(self, manager):
         self._manager = weakref.ref(manager)  # Don't create reference cycles
 
         # Slight misnomer. Only locks out changed events. Doesn't actually prevent other changes.
-        self._change_lock = threading.Lock()
+        self._change_lock = threading.Lock()  # XXX: Use asyncio.Lock? Only if .update() is a coroutine
 
         self._props = {}
 
@@ -149,6 +97,7 @@ class Block(dbus.service.Object):
         Performs a bulk update of properties. Issues only one changed event for
         the entire update.
         """
+        # XXX: Rewrite as aio coroutine?
         with self._change_lock:
             for prop, value in values.items():
                 if prop in self._WELL_KNOWN_PROPS:
@@ -161,7 +110,7 @@ class Block(dbus.service.Object):
         """
         Raise a changed event, unless it's been locked out by update()
         """
-        have_lock = self._change_lock.acquire()
+        have_lock = self._change_lock.acquire(blocking=False)
         if have_lock:
             self._change_lock.release()
             self.changed()
@@ -420,22 +369,3 @@ class Block(dbus.service.Object):
     def set_order(self, value):
         self._order = value
         self._changed()
-
-
-def start():
-    DBusGMainLoop(set_as_default=True)
-
-    try:
-        with open("{}/.i3-dstatus.conf".format(os.path.expanduser('~'))) as f:
-            config = yaml.safe_load(f)
-    except FileNotFoundError:
-        config = {}
-
-    manager = BlockManager(config)
-
-    sys.stdout.write('{"version":1}\n[\n[]\n')
-    # sys.stdout.write('{"version":1, "click_events":true}\n[\n[]\n')
-    sys.stdout.flush()
-
-    main = GLib.MainLoop()
-    main.run()
